refactor(editor): add rich text editor with toolbar and markdown support
- Implement EditorToolbar component with formatting options including
  bold, italic, strikethrough, code, headings, lists, and blockquotes
- Add ToolbarButton component for consistent toolbar styling
- Create helper functions for markdown operations including
  getMarkdownFromEditor, setMarkdownToEditor, and applyMarkdownSyntax
- Integrate editor toolbar in message input component with top/bottom
  positioning options
- Support source mode toggle between visual and markdown editing
- Add file attachment functionality with proper markdown generation
- Enhance editor styling with improved UI and responsive design
- Add expanded modal view for full markdown editing experience
- Include proper internationalization support for editor features

The new editor provides comprehensive text formatting capabilities while
maintaining compatibility with markdown syntax for content storage.
