+++ justfile
@@ -81,7 +81,7 @@ clean:
     rm app.db
     rm app.db-shm
     rm app.db-wal
-    
+
 
 # ── 别名 ─────────────────────────────────────────────
